add operator stubs: o/operator_stub/%.o will be compiled with MAPPING_OPERATOR_STUBS=1 When an operator is compiled as a stub, it should not do anything but parsing the parameters and implementing writeSemanticParameters(). The get*()-methods must be omitted, and external dependencies should be kept to a minimum.
Stubs are meant to be compiled into the cgi or indexserver so they can verify and/or canonicalize a query string without containing the actual processing logic.

In this patchset, several operators had their writeSemanticParameters() method added or fixed. The timeshift operator was removed from compilation until it can implement a proper writeSemanticParameters().