Added a systemtest for Points2Histogram operator.
Therefore I needed to extend the mapping_manager to cope with plots.

Then I needed to implement the `hash` function in plot, which uses toJSON.
I changed toJSON in this sense to be const and return a const string (as long as there is no mutable string necessary anywhere).

XYPlot now makes it necessary to sort it before calling toJSON.
As this is done in the operator and checked as well it should not cause any problems.

Furthermore I needed to implement the clone method for Histogram.
As this is different for each plot type, I made it purely virtual.
The implementation for Text and Png is still incomplete (but will fail on any systemtest for this operator).

Bugfix: Changed name of time_shift systemtest.