remove template from invocations of buffer->template read.
The problem seems to be that write_and_read() is a template, and if the result of that template call is assigned to a auto variable, clang isn't too sure about its type. Considering that unique_ptr is also a template, name lookup seems to fail.

The workaround is to define the variable as the proper type instead of using auto.